+++ db/src/main/scala/serv1/model/HistoricalData.scala
@@ -4,10 +4,10 @@ import serv1.db.types.HistoricalDataType.HistoricalDataType
 
 case class HistoricalData(
                            var timestamp: Long,
-                           var high: Long,
-                           var low: Long,
-                           var open: Long,
-                           var close: Long,
+                           var high: Double,
+                           var low: Double,
+                           var open: Double,
+                           var close: Double,
                            var vol: Double,
                            var historicalDataType: HistoricalDataType
                          )
